refactor/#65: 오타 수정
